JBIDE-25556 - api cleanup and splitting utils out
