+++ common/plugins/org.jboss.tools.common.jdt.debug/src/org/jboss/tools/common/jdt/debug/tools/internal/Tools.java
@@ -11,36 +11,23 @@
 package org.jboss.tools.common.jdt.debug.tools.internal;
 
 import java.io.File;
-import java.io.IOException;
 import java.io.InputStream;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Method;
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.net.URLClassLoader;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.List;
 import java.util.Properties;
 import java.util.Set;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 import org.eclipse.core.runtime.IStatus;
-import org.eclipse.core.runtime.Platform;
 import org.eclipse.core.runtime.Preferences.IPropertyChangeListener;
 import org.eclipse.core.runtime.Preferences.PropertyChangeEvent;
 import org.eclipse.core.runtime.preferences.IEclipsePreferences;
 import org.eclipse.core.runtime.preferences.IEclipsePreferences.IPreferenceChangeListener;
 import org.eclipse.core.runtime.preferences.IEclipsePreferences.PreferenceChangeEvent;
 import org.eclipse.core.runtime.preferences.InstanceScope;
-import org.eclipse.jdt.launching.IVMInstall;
-import org.eclipse.jdt.launching.IVMInstall2;
-import org.eclipse.jdt.launching.IVMInstallType;
 import org.eclipse.jdt.launching.JavaRuntime;
 import org.eclipse.osgi.util.NLS;
 import org.jboss.tools.common.jdt.debug.IPropertyKeys;
+import org.jboss.tools.common.jdt.debug.JavaUtilities;
 import org.jboss.tools.common.jdt.debug.Messages;
 import org.jboss.tools.common.jdt.debug.RemoteDebugActivator;
 import org.jboss.tools.common.jdt.debug.tools.ToolsCoreException;
@@ -109,19 +96,6 @@ public class Tools implements IPreferenceChangeListener, IToolsConstants, IPrope
 		return toolsJarFile;
 	}
 
-	private boolean isJigsawRunning() {
-		IVMInstall vmi = findActiveVM();
-		if( vmi instanceof IVMInstall2 ) {
-			IVMInstall2 vmi2 = (IVMInstall2)vmi;
-			String javaV = vmi2.getJavaVersion();
-			if( javaV != null && javaV.startsWith("9.")) {
-				// TODO add support for java10 etc 
-				return true;
-			}
-		}
-		return false;
-	}
-	
 	public void reset() {
 		toolsLoader = null;
 		getToolsLoader();
@@ -131,39 +105,17 @@ public class Tools implements IPreferenceChangeListener, IToolsConstants, IPrope
 	private ClassLoader getToolsLoader() {
 		if (toolsLoader == null) {
 			toolsLoader = Thread.currentThread().getContextClassLoader();
-			
-			if( !isJigsawRunning()) {
 
-				File toolsJar = getToolsJar();
-				if (toolsJar == null) {
-					return toolsLoader;
+			// tools.jar classes are available directly on classpath in java9 and above.
+			if( !JavaUtilities.isJigsawRunning()) {
+				PreJigsawUtility util = new PreJigsawUtility();
+				File toolsJar = util.getToolsJarFile();
+				if( toolsJar != null ) {
+					toolsLoader = util.getToolsJarClassloader(toolsJar, toolsLoader);
 				}
-				try {
-					toolsJar = toolsJar.getCanonicalFile();
-				} catch (IOException e1) {
-					return toolsLoader;
-				}
-				if (toolsJar.exists()) {
-					toolsJarFile = toolsJar;
-					URL toolsUrl;
-					try {
-						toolsUrl = toolsJar.toURI().toURL();
-					} catch (MalformedURLException e) {
-						return toolsLoader;
-					}
-					toolsLoader = new URLClassLoader(new URL[] { toolsUrl }, toolsLoader);
-				}
-				
 			}
 		}
 		return toolsLoader;
-	}
-
-	private File getToolsJar() {
-		// Find the jdk root for the currently running java home
-		String jdkRootDirectory = findHomeDirectoryToAddToClasspath();
-		File file = new File(jdkRootDirectory + TOOLS_JAR);
-		return file;
 	}
 
 	/**
@@ -173,203 +125,6 @@ public class Tools implements IPreferenceChangeListener, IToolsConstants, IPrope
 	 */
 	public boolean isReady() {
 		return isReady;
-	}
-
-	/**
-	 * Validates the JDK root directory.
-	 * 
-	 * @param jdkRootDirectory
-	 *            The JDK root directory
-	 * @return The error message or <tt>null</tt> if not found
-	 */
-	public String validateJdkRootDirectory(String jdkRootDirectory) {
-		if (jdkRootDirectory == null || jdkRootDirectory.trim().isEmpty()) {
-			return Messages.NoJdkDirectoryFoundMsg;
-		}
-		// check if directory exists
-		File directory = new File(jdkRootDirectory);
-		if (!directory.exists() || !directory.isDirectory()) {
-			return Messages.directoryNotExistMsg;
-		}
-
-		// check if tools.jar exists
-		File toolsJarFile = new File(jdkRootDirectory + TOOLS_JAR);
-		if (!toolsJarFile.exists()) {
-			return Messages.notJdkRootDirectoryMsg;
-		}
-
-		// checks if "attach" shared library exist
-		String libraryPath = getJreLibraryPath(jdkRootDirectory);
-		if (libraryPath == null) {
-			return Messages.notJdkRootDirectoryMsg;
-		}
-		return null;
-	}
-
-	private IVMInstall[] getAllVMInstalls() {
-		ArrayList<IVMInstall> all = new ArrayList<IVMInstall>();
-		for (IVMInstallType type : JavaRuntime.getVMInstallTypes()) {
-			for (IVMInstall install : type.getVMInstalls()) {
-				all.add(install);
-			}
-		}
-		return (IVMInstall[]) all.toArray(new IVMInstall[all.size()]);
-	}
-
-	private int compareJavaVersions(String version0, String version1) {
-		int[] arg0majorMinor = getMajorMinor(version0);
-		int[] arg1majorMinor = getMajorMinor(version1);
-		if (arg0majorMinor[0] < arg1majorMinor[0])
-			return -1;
-		if (arg0majorMinor[0] > arg1majorMinor[0])
-			return 1;
-		if (arg0majorMinor[1] < arg1majorMinor[1])
-			return -1;
-		if (arg0majorMinor[1] > arg1majorMinor[1])
-			return 1;
-		return 0;
-	}
-
-	/**
-	 * Get all IVMInstalls that have a major.minor <= the currently running
-	 * jre's major.minor
-	 * 
-	 * @return
-	 */
-	public IVMInstall[] getAllCompatibleInstalls() {
-		String currentVersion = System.getProperty("java.version");
-		ArrayList<IVMInstall2> compat = new ArrayList<IVMInstall2>();
-		IVMInstall[] all = getAllVMInstalls();
-		for (int i = 0; i < all.length; i++) {
-			if (all[i] instanceof IVMInstall2) {
-				String vers = ((IVMInstall2) all[i]).getJavaVersion();
-				// if running jre is greater than or equal the vm install, 
-				// it's probably compatible
-				if (compareJavaVersions(currentVersion, vers) >= 0) {
-					compat.add((IVMInstall2) all[i]);
-				}
-			}
-		}
-
-		// Sort them by version number, so higher matches are at the beginning
-		// of the list
-		Comparator<IVMInstall2> comparator = new Comparator<IVMInstall2>() {
-			public int compare(IVMInstall2 arg0, IVMInstall2 arg1) {
-				String arg0vers = arg0.getJavaVersion();
-				String arg1vers = arg1.getJavaVersion();
-				return compareJavaVersions(arg0vers, arg1vers);
-			}
-		};
-
-		Collections.sort(compat, comparator);
-		Collections.reverse(compat);
-		return (IVMInstall[]) compat.toArray(new IVMInstall[compat.size()]);
-	}
-
-	private int[] getMajorMinor(String version) {
-		Matcher m = Pattern.compile("^(\\d+)\\.(\\d+)\\..*").matcher(version);
-		if (!m.matches()) {
-			throw new IllegalArgumentException("Malformed version string");
-		}
-		return new int[] { Integer.parseInt(m.group(1)), // major
-				Integer.parseInt(m.group(2)) };
-	}
-
-	/**
-	 * Find the first vm-install that is valid
-	 * 
-	 * @return The JDK root directory, or empty string if not found
-	 */
-	private IVMInstall findFirstValidVMInstall() {
-		// search from the JREs that are compatible based on java version
-		IVMInstall[] all = getAllCompatibleInstalls();
-		for (int i = 0; i < all.length; i++) {
-			String jdkRootDirectory = all[i].getInstallLocation().getPath();
-			if (null == validateJdkRootDirectory(jdkRootDirectory)) {
-				return all[i];
-			}
-		}
-		RemoteDebugActivator.pluginLog().logMessage(IStatus.WARNING, Messages.jdkRootDirectoryNotFoundMsg,
-				new Exception(Messages.jdkRootDirectoryNotFoundMsg));
-		return null; 
-	}
-
-	public String findJdkRootFromJavaHome() {
-		// search at the same directory as current JRE
-		String javaHome = System.getProperty(JAVA_HOME_PROPERTY_KEY);
-		for (File directory : getPossibleJdkRootDirectory(javaHome)) {
-			String path = directory.getPath();
-			if (null == validateJdkRootDirectory(path)) {
-				//RemoteDebugActivator.pluginLog().logInfo(NLS.bind(Messages.jdkRootDirectoryFoundMsg, path));
-				return path;
-			}
-		}
-		return null;
-	}
-
-	/**
-	 * Gets the directories that could be JDK root directory.
-	 * 
-	 * @param javaHome
-	 *            The java home path
-	 * @return The directories that could be JDK root directory.
-	 */
-	private static List<File> getPossibleJdkRootDirectory(String javaHome) {
-		List<File> dirs = new ArrayList<File>();
-
-		/*
-		 * On Mac, java home path can be for example:
-		 * /Library/Java/JavaVirtualMachines/jdk1.7.0_13.jdk/Contents/Home/jre
-		 */
-
-		if (Platform.getOS().equals(Platform.OS_MACOSX)) {
-			int index = javaHome.indexOf(IToolsConstants.JAVA_INSTALLATION_DIR_ON_MAC);
-			if (index == -1) {
-				return dirs;
-			}
-
-			String javaVirtualMachinesPath = javaHome.substring(0,
-					index + IToolsConstants.JAVA_INSTALLATION_DIR_ON_MAC.length());
-			File dir = new File(javaVirtualMachinesPath);
-
-			collectDirs(dirs, dir, 3);
-			return dirs;
-		}
-
-		File parentDir = new File(javaHome + File.separator + ".."); //$NON-NLS-1$
-		// JRE's most often live inside a jdk's jre folder
-		dirs.add(parentDir);
-		if (parentDir.exists()) {
-			for (File file : parentDir.listFiles()) {
-				if (file.isDirectory()) {
-					dirs.add(file);
-				}
-			}
-		}
-
-		return dirs;
-	}
-
-	/**
-	 * Collects the directories which are within given depth from given base
-	 * directory.
-	 * 
-	 * @param dirs
-	 *            The directories to store result
-	 * @param dir
-	 *            The directory to search
-	 * @param depth
-	 *            The depth to search
-	 */
-	private static void collectDirs(List<File> dirs, File dir, int depth) {
-		if (depth > 0) {
-			for (File file : dir.listFiles()) {
-				if (file.isDirectory()) {
-					dirs.add(file);
-					collectDirs(dirs, file, depth - 1);
-				}
-			}
-		}
 	}
 
 	/**
@@ -385,110 +140,6 @@ public class Tools implements IPreferenceChangeListener, IToolsConstants, IPrope
 			return false;
 		}
 		return true;
-	}
-
-	// Find the home directory based on the preference settings
-	private IVMInstall findHomeDirFromPreferences() {
-		IEclipsePreferences prefs = InstanceScope.INSTANCE.getNode(RemoteDebugActivator.PLUGIN_ID);
-		String vm = prefs.get(IPropertyKeys.JDK_VM_INSTALL, null);
-		if (vm != null) {
-			IVMInstall found = findVMInstall(vm);
-			if (found != null) {
-				return found;
-			}
-		}
-		return null;
-	}
-
-	private static IVMInstall findVMInstall(String id) {
-		if (id != null) {
-			IVMInstallType[] types = JavaRuntime.getVMInstallTypes();
-			IVMInstall ret = null;
-			for (int i = 0; i < types.length; i++) {
-				ret = types[i].findVMInstall(id);
-				if (ret != null)
-					return ret;
-			}
-		}
-		return null;
-	}
-
-	private static IVMInstall findVMByInstallLocation(String home) {
-		if (home != null) {
-			IVMInstallType[] types = JavaRuntime.getVMInstallTypes();
-			for (int i = 0; i < types.length; i++) {
-				IVMInstall[] installs = types[i].getVMInstalls();
-				for (int j = 0; j < installs.length; j++) {
-					File installLoc = installs[j].getInstallLocation();
-					if (home.equals(installLoc.getAbsolutePath())) {
-						return installs[j];
-					}
-				}
-			}
-		}
-		return null;
-	}
-
-	private String findHomeDirectoryToAddToClasspath() {
-		// Find the jdk root for the currently running java home
-		String jdkRootDirectory = findJdkRootFromJavaHome();
-		if (jdkRootDirectory == null || jdkRootDirectory.trim().isEmpty()) {
-			IVMInstall vmi = findSecondaryVMInstall();
-			if (vmi != null)
-				jdkRootDirectory = vmi.getInstallLocation().getAbsolutePath();
-		}
-		return jdkRootDirectory;
-	}
-
-	/**
-	 * Get the vminstall to use when the currently-running jre is not a jdk
-	 * 
-	 * @return
-	 */
-	public IVMInstall findSecondaryVMInstall() {
-		IVMInstall vmi = findHomeDirFromPreferences();
-		if (vmi == null) {
-			vmi = findFirstValidVMInstall();
-		}
-		return vmi;
-	}
-
-	/**
-	 * Find the IVMInstall who's home directory matches the current java.home
-	 * sysprop
-	 * 
-	 * @return
-	 */
-	public IVMInstall findActiveVM() {
-		String javaHome = System.getProperty(JAVA_HOME_PROPERTY_KEY);
-		return findVMByInstallLocation(javaHome);
-	}
-
-	/**
-	 * Does an IVMInstall currently exist for the currently-running java.home
-	 * 
-	 * @return
-	 */
-	public boolean hasActiveVMInstall() {
-		return findActiveVM() != null;
-	}
-
-	/**
-	 * Gets the JRE library path.
-	 * 
-	 * @param jdkRootDirectory
-	 *            The JDK root directory
-	 * @return The JRE library path or <tt>null</tt> it not found
-	 */
-	private static String getJreLibraryPath(String jdkRootDirectory) {
-		for (String path : LIBRARY_PATHS) {
-			File attachLibraryFile = new File(jdkRootDirectory + path + File.separator
-					+ System.mapLibraryName(ATTACH_LIBRARY));
-			if (attachLibraryFile.exists()) {
-				return jdkRootDirectory + path;
-			}
-		}
-		return null;
 	}
 
 	/*
